script output fix
