+++ Photon.Agent/MessageHandlers/BuildSessionReleaseProcessor.cs
@@ -2,6 +2,7 @@
 using Photon.Communication;
 using Photon.Library.Messages;
 using System.Threading.Tasks;
+using Photon.Framework.Messages;
 
 namespace Photon.Agent.MessageHandlers
 {
