+++ Photon.Framework/Photon.Framework.csproj
@@ -56,6 +56,10 @@
     <Compile Include="Extensions\StringExtensions.cs" />
     <Compile Include="Extensions\RemoteTaskExtensions.cs" />
     <Compile Include="FileUtils.cs" />
+    <Compile Include="Messages\BuildSessionBeginRequest.cs" />
+    <Compile Include="Messages\BuildSessionBeginResponse.cs" />
+    <Compile Include="Messages\BuildSessionReleaseRequest.cs" />
+    <Compile Include="Messages\BuildSessionReleaseResponse.cs" />
     <Compile Include="NetPath.cs" />
     <Compile Include="Packages\FilePatternMatching.cs" />
     <Compile Include="Packages\PackageFileDefinition.cs" />
@@ -65,6 +69,8 @@
     <Compile Include="Packages\ProjectPackage.cs" />
     <Compile Include="Domain\RemoteTaskCompletionSource.cs" />
     <Compile Include="Projects\Project.cs" />
+    <Compile Include="Scripts\AgentBuildSessionHandle.cs" />
+    <Compile Include="Scripts\AgentDeploySessionHandle.cs" />
     <Compile Include="Scripts\DeployScriptRegistry.cs" />
     <Compile Include="Scripts\IAgentSessionHandle.cs" />
     <Compile Include="Scripts\IDeployScript.cs" />
@@ -76,6 +82,8 @@
     <Compile Include="Scripts\ScriptOutput.cs" />
     <Compile Include="Scripts\BuildScriptRegistry.cs" />
     <Compile Include="Scripts\ScriptResult.cs" />
+    <Compile Include="Scripts\ServerBuildContext.cs" />
+    <Compile Include="Scripts\ServerDeployContext.cs" />
     <Compile Include="ServerAgentDefinition.cs" />
     <Compile Include="ServerDefinition.cs" />
     <Compile Include="ServerHttpDefinition.cs" />
@@ -99,6 +107,12 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Sessions\" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Photon.Communication\Photon.Communication.csproj">
+      <Project>{E167A349-7691-4406-9F4B-989BD5E6E12F}</Project>
+      <Name>Photon.Communication</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
