+++ Photon.Server/HttpHandlers/BuildHandler.cs
@@ -1,13 +1,14 @@
 ﻿using log4net;
 using Newtonsoft.Json;
 using Photon.Framework.Extensions;
+using Photon.Framework.Scripts;
 using Photon.Library.Messages;
 using Photon.Server.Internal;
-using Photon.Server.Internal.Scripts;
 using Photon.Server.Internal.Sessions;
 using PiServerLite.Http.Handlers;
 using System;
 using System.IO;
+using Photon.Framework.Messages;
 
 namespace Photon.Server.HttpHandlers
 {
