+++ PhotonCLI/Actions/BuildRunAction.cs
@@ -9,6 +9,7 @@ using System;
 using System.IO;
 using System.Net;
 using System.Threading.Tasks;
+using Photon.Framework.Messages;
 
 namespace Photon.CLI.Actions
 {
